checker/requirements.txt
aiogram==3.13.1
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-auth==2.35.0
google-api-python-client==2.149.0
